Started Work on FoV cone for enemies
In the TacHud UI listener there is the code for FoV, most of it is copy
paste from the original code (which i fear we'd have to override the
classes for because Event Listener returns dont really work) the actual
FoV code is working but there are some bugs with the original code still
running. The FoV code is all the vector math in the middle there (sorry
that there's no doc. I normally write it in the end)