Implemented Concealment effects
A few bugs are still in the system. For some reason GC is either not
working or the code is bugged with generation of tiles since it still
has concealment effects even after the enemies have left the position or
concealment was broken. the floating red eye effects are not going away
for some reason so i cant rebuild the effects. Pushing to GitHub,
documentation? sometime soon when i finish with this